attempt to fix path to federated_instances, which changed after the release of v0.18.0
 * https://github.com/maltfield/awesome-lemmy-instances/issues/30

At first I thought I'd have to query the now-segregated federated_instances endpoint myself, but now I realize that it was just moved to a different path of the JSON file. Curiously, it's now at the root as 'federated_instances' and inside of that is another 'federated_instances' dictionary. Strange, but it appears to work now.